Rename validateMemos.ts->memoUtils.ts
