+++ src/components/scenes/TransactionDetailsScene.tsx
@@ -22,8 +22,8 @@ import { useSelector } from '../../types/reactRedux'
 import { EdgeSceneProps } from '../../types/routerTypes'
 import { getCurrencyCodeWithAccount } from '../../util/CurrencyInfoHelpers'
 import { matchJson } from '../../util/matchJson'
+import { getMemoTitle } from '../../util/memoUtils'
 import { convertCurrencyFromExchangeRates, convertNativeToExchange, darkenHexColor, removeIsoPrefix } from '../../util/utils'
-import { getMemoTitle } from '../../util/validateMemos'
 import { ButtonsView } from '../buttons/ButtonsView'
 import { AdvancedDetailsCard } from '../cards/AdvancedDetailsCard'
 import { EdgeCard } from '../cards/EdgeCard'
